requirements-lambda.txt
# Core FastAPI and ASGI
fastapi==0.116.1
starlette==0.41.3
uvicorn[standard]==0.35.0
mangum==0.19.0

# Database (SQLite for Lambda)
sqlalchemy==2.0.43
aiosqlite==0.20.0

# HTTP client
httpx==0.28.1

# Form handling
python-multipart==0.0.20

# Configuration
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1

# AWS and other services
boto3==1.34.0
twilio==9.8.0

# Date/time processing
parsedatetime==2.6
python-dateutil==2.8.2

# Phone number handling
phonenumbers==8.13.50

# Security
itsdangerous==2.2.0

# Logging
structlog==23.2.0

# Required dependencies
typing_extensions==4.15.0
pydantic-core==2.32.0
annotated-types==0.7.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
anyio==3.7.1
sniffio==1.3.1
h11==0.14.0
httpcore==1.0.8
certifi==2024.12.14
click==8.3.0
colorama==0.4.6
greenlet==3.1.1